requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.7
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.2
pluggy==1.5.0
pymango==0.1.1
pymongo==4.10.1
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
tomli==2.1.0
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
